Update AttackPowerUp.cs
updated the duration of how long the attack will last on the screen to 1 second.
Changed it to having the attack gameobject active only if the attacking boolean is true.
Removed the animation trigger from here to the player controller script instead and created 2 methods that allow for getting and setting the attacking bool.