Added sound effects into the game
Created an audio source that will play various sound effects in the game.
Audio controller is used to control which sound effect/audio shall be played.
Added various sound effect files for use in the game along with having them play when the player jumps/attacks or collected the power up item.